Registration opened and workshops
Changed to be a link instead of embedded iframe, which was not responsive. There is commented out code that will be removed after this years' event. I also attempted to have the grid.js refresh when the spanish checkbox is true, but it was outside my comfort area. The grid is not that great for mobile, so will be modifying the display of workshops for the next event.