Open Registration
Changed embedded iframe to a link as google form isn't responsive enough. There are also some changes where I was trying to get the grid.js to refresh when LANGUAGE==spanish. I plan to remove the grid and replace it with similar to woking like the rest of the content. The hover over isn't that great of a mobile experience.